fix(reader): rename layout helper so Next does not treat it as a route layout
Any layout.ts under src/app is parsed as a route segment layout, so the
helper was type checked against LayoutConfig<'/reader/utils'> and failed
the web build. pnpm lint does not run that check; only next build does.
