+++ apps/readest-app/src/app/reader/utils/mobileLayout.ts
